+++ src/xmp.cpp
@@ -383,14 +383,14 @@ namespace Exiv2 {
         return 0;
     }
 
-    long Xmpdatum::count() const
+    size_t Xmpdatum::count() const
     {
         return p_->value_.get() == nullptr ? 0 : p_->value_->count();
     }
 
     long Xmpdatum::size() const
     {
-        return p_->value_.get() == nullptr ? 0 : p_->value_->size();
+        return p_->value_.get() == nullptr ? 0 : static_cast<long>(p_->value_->size());
     }
 
     std::string Xmpdatum::toString() const
@@ -951,7 +951,7 @@ namespace Exiv2 {
             if (i.typeId() == xmpBag || i.typeId() == xmpSeq || i.typeId() == xmpAlt) {
                 printNode(ns, i.tagName(), "", options);
                 meta.SetProperty(ns.c_str(), i.tagName().c_str(), nullptr, options);
-                for (long idx = 0; idx < i.count(); ++idx) {
+                for (size_t idx = 0; idx < i.count(); ++idx) {
                     const std::string item = i.tagName() + "[" + toString(idx + 1) + "]";
                     printNode(ns, item, i.toString(idx), 0);
                     meta.SetProperty(ns.c_str(), item.c_str(), i.toString(idx).c_str());
